Went back to using the mpu library :(
I thought that the acc only measures gravity, but unfortunately it measures every linear acc caused
by external forces. For our scenario, this might not be a problem, since the sensor is located
right at the center of rotation. But for a future project (eg. drone), we need the library (Sensor Fusion capabilities).
So, what changed:
	- minor general fixes to the code
	- use of sensor library
	- i kept DLPF setting